MorphoTrening/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
pymorphy3>=2.0.0
pdfminer.six>=20221105
